[XABT] Move additional provider source code generation out of `<GenerateMainAndroidManifest>` task.
