+++ src/Xamarin.Android.Build.Tasks/Tasks/GenerateMainAndroidManifest.cs
@@ -1,14 +1,10 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using Java.Interop.Tools.Cecil;
-using Java.Interop.Tools.JavaCallableWrappers;
-using Java.Interop.Tools.TypeNameMappings;
 using Microsoft.Android.Build.Tasks;
 using Microsoft.Build.Framework;
-using Mono.Cecil;
+using Microsoft.Build.Utilities;
 using Xamarin.Android.Tools;
 
 namespace Xamarin.Android.Tasks;
@@ -17,6 +13,8 @@ public class GenerateMainAndroidManifest : AndroidTask
 {
 	public override string TaskPrefix => "GMM";
 
+	[Output]
+	public ITaskItem []? AdditionalProviderSources { get; set; }
 	[Required]
 	public string AndroidRuntime { get; set; } = "";
 	public string? AndroidSdkDir { get; set; }
@@ -25,10 +23,8 @@ public class GenerateMainAndroidManifest : AndroidTask
 	public string? ApplicationLabel { get; set; }
 	public string? BundledWearApplicationName { get; set; }
 	public string? CheckedBuild { get; set; }
-	public string CodeGenerationTarget { get; set; } = "";
 	public bool Debug { get; set; }
 	public bool EmbedAssemblies { get; set; }
-	public bool EnableMarshalMethods { get; set; }
 	[Required]
 	public string IntermediateOutputDirectory { get; set; } = "";
 	public string []? ManifestPlaceholders { get; set; }
@@ -37,22 +33,16 @@ public class GenerateMainAndroidManifest : AndroidTask
 	public string []? MergedManifestDocuments { get; set; }
 	public bool MultiDex { get; set; }
 	public bool NeedsInternet { get; set; }
-	public string? OutputDirectory { get; set; }
 	public string? PackageName { get; set; }
 	[Required]
 	public ITaskItem [] ResolvedUserAssemblies { get; set; } = [];
 	[Required]
 	public string [] SupportedAbis { get; set; } = [];
 	public string? SupportedOSPlatformVersion { get; set; }
-	[Required]
-	public string TargetName { get; set; } = "";
 	public string? VersionCode { get; set; }
 	public string? VersionName { get; set; }
 
 	AndroidRuntime androidRuntime;
-	JavaPeerStyle codeGenerationTarget;
-
-	bool UseMarshalMethods => !Debug && EnableMarshalMethods;
 
 	public override bool RunTask ()
 	{
@@ -68,21 +58,18 @@ public class GenerateMainAndroidManifest : AndroidTask
 		var userAssembliesPerArch = MonoAndroidHelper.GetPerArchAssemblies (ResolvedUserAssemblies, SupportedAbis, validate: true);
 
 		androidRuntime = MonoAndroidHelper.ParseAndroidRuntime (AndroidRuntime);
-		codeGenerationTarget = MonoAndroidHelper.ParseCodeGenerationTarget (CodeGenerationTarget);
 
 		// Generate the merged manifest
 		var additionalProviders = MergeManifest (templateCodeGenState, GenerateJavaStubs.MaybeGetArchAssemblies (userAssembliesPerArch, templateCodeGenState.TargetArch));
-		GenerateAdditionalProviderSources (templateCodeGenState, additionalProviders);
 
+		AdditionalProviderSources = additionalProviders.Select (p => new TaskItem (p)).ToArray ();
 
-		// If we still need the NativeCodeGenState in the <GenerateNativeMarshalMethodSources> task because we're using marshal methods,
-		// we're going to transfer it to a new object that doesn't require holding open Cecil AssemblyDefinitions.
-		if (UseMarshalMethods) {
-			var nativeCodeGenStateObject = MarshalMethodCecilAdapter.GetNativeCodeGenStateCollection (Log, nativeCodeGenStates);
+		// We still need the NativeCodeGenState for later tasks, but we're going to transfer
+		// it to a new object that doesn't require holding open Cecil AssemblyDefinitions.
+		var nativeCodeGenStateObject = MarshalMethodCecilAdapter.GetNativeCodeGenStateCollection (Log, nativeCodeGenStates);
 
-			Log.LogDebugMessage ($"Saving {nameof (NativeCodeGenStateObject)} to {nameof (GenerateJavaStubs.NativeCodeGenStateObjectRegisterTaskKey)}");
-			BuildEngine4.RegisterTaskObjectAssemblyLocal (MonoAndroidHelper.GetProjectBuildSpecificTaskObjectKey (GenerateJavaStubs.NativeCodeGenStateObjectRegisterTaskKey, WorkingDirectory, IntermediateOutputDirectory), nativeCodeGenStateObject, RegisteredTaskObjectLifetime.Build);
-		}
+		Log.LogDebugMessage ($"Saving {nameof (NativeCodeGenStateObject)} to {nameof (GenerateJavaStubs.NativeCodeGenStateObjectRegisterTaskKey)}");
+		BuildEngine4.RegisterTaskObjectAssemblyLocal (MonoAndroidHelper.GetProjectBuildSpecificTaskObjectKey (GenerateJavaStubs.NativeCodeGenStateObjectRegisterTaskKey, WorkingDirectory, IntermediateOutputDirectory), nativeCodeGenStateObject, RegisteredTaskObjectLifetime.Build);
 
 		// Dispose the Cecil resolvers so the assemblies are closed.
 		Log.LogDebugMessage ($"Disposing all {nameof (NativeCodeGenState)}.{nameof (NativeCodeGenState.Resolver)}");
@@ -139,81 +126,5 @@ public class GenerateMainAndroidManifest : AndroidTask
 		}
 
 		return additionalProviders;
-	}
-
-	void GenerateAdditionalProviderSources (NativeCodeGenState codeGenState, IList<string> additionalProviders)
-	{
-		// Create additional runtime provider java sources.
-		bool isMonoVM = androidRuntime switch {
-			Xamarin.Android.Tasks.AndroidRuntime.MonoVM => true,
-			Xamarin.Android.Tasks.AndroidRuntime.CoreCLR => true,
-			_ => false,
-		};
-		string providerTemplateFile = isMonoVM ?
-			"MonoRuntimeProvider.Bundled.java" :
-			"NativeAotRuntimeProvider.java";
-		string providerTemplate = GetResource (providerTemplateFile);
-
-		foreach (var provider in additionalProviders) {
-			var contents = providerTemplate.Replace (isMonoVM ? "MonoRuntimeProvider" : "NativeAotRuntimeProvider", provider);
-			var real_provider = isMonoVM ?
-				Path.Combine (OutputDirectory, "src", "mono", provider + ".java") :
-				Path.Combine (OutputDirectory, "src", "net", "dot", "jni", "nativeaot", provider + ".java");
-			Files.CopyIfStringChanged (contents, real_provider);
-		}
-
-		// For NativeAOT, generate JavaInteropRuntime.java
-		if (androidRuntime == Xamarin.Android.Tasks.AndroidRuntime.NativeAOT) {
-			const string fileName = "JavaInteropRuntime.java";
-			string template = GetResource (fileName);
-			var contents = template.Replace ("@MAIN_ASSEMBLY_NAME@", TargetName);
-			var path = Path.Combine (OutputDirectory, "src", "net", "dot", "jni", "nativeaot", fileName);
-			Log.LogDebugMessage ($"Writing: {path}");
-			Files.CopyIfStringChanged (contents, path);
-		}
-
-		// Create additional application java sources.
-		StringWriter regCallsWriter = new StringWriter ();
-		regCallsWriter.WriteLine ("// Application and Instrumentation ACWs must be registered first.");
-		foreach (TypeDefinition type in codeGenState.JavaTypesForJCW) {
-			if (JavaNativeTypeManager.IsApplication (type, codeGenState.TypeCache) || JavaNativeTypeManager.IsInstrumentation (type, codeGenState.TypeCache)) {
-				if (codeGenState.Classifier != null && !codeGenState.Classifier.TypeHasDynamicallyRegisteredMethods (type)) {
-					continue;
-				}
-
-				string javaKey = JavaNativeTypeManager.ToJniName (type, codeGenState.TypeCache).Replace ('/', '.');
-				regCallsWriter.WriteLine (
-					codeGenerationTarget == JavaPeerStyle.XAJavaInterop1 ?
-						"\t\tmono.android.Runtime.register (\"{0}\", {1}.class, {1}.__md_methods);" :
-						"\t\tnet.dot.jni.ManagedPeer.registerNativeMembers ({1}.class, {1}.__md_methods);",
-					type.GetAssemblyQualifiedName (codeGenState.TypeCache),
-					javaKey
-				);
-			}
-		}
-		regCallsWriter.Close ();
-
-		var real_app_dir = Path.Combine (OutputDirectory, "src", "net", "dot", "android");
-		string applicationTemplateFile = "ApplicationRegistration.java";
-		SaveResource (
-			applicationTemplateFile,
-			applicationTemplateFile,
-			real_app_dir,
-			template => template.Replace ("// REGISTER_APPLICATION_AND_INSTRUMENTATION_CLASSES_HERE", regCallsWriter.ToString ())
-		);
-	}
-
-	string GetResource (string resource)
-	{
-		using (var stream = GetType ().Assembly.GetManifestResourceStream (resource))
-		using (var reader = new StreamReader (stream))
-			return reader.ReadToEnd ();
-	}
-
-	void SaveResource (string resource, string filename, string destDir, Func<string, string> applyTemplate)
-	{
-		string template = GetResource (resource);
-		template = applyTemplate (template);
-		Files.CopyIfStringChanged (template, Path.Combine (destDir, filename));
 	}
 }
